+++ compose.yaml
@@ -57,8 +57,8 @@ services:
 
   dashboard:
     build:
-      context: ./dashboard
-      dockerfile: ./Dockerfile
+      context: .
+      dockerfile: ./dashboard/Dockerfile
     image: dashboard:latest
     environment:
       ENVIRONMENT: ${ENVIRONMENT:-dev}
