+++ dashboard/Dockerfile
@@ -1,6 +1,6 @@
 FROM ghcr.io/astral-sh/uv:python3.12-bookworm-slim
 
-# Install build dependencies (keep these if needed for other packages)
+# Install build dependencies
 RUN apt-get update && apt-get install -y \
     build-essential \
     gcc \
@@ -11,11 +11,9 @@ RUN apt-get update && apt-get install -y \
     && rm -rf /var/lib/apt/lists/*
 
 WORKDIR /app
-COPY pyproject.toml uv.lock ./
 
 ENV UV_COMPILE_BYTECODE=1
 ENV UV_LINK_MODE=copy
-# Set piwheels as an extra index for package resolution
 ENV UV_EXTRA_INDEX_URL=https://www.piwheels.org/simple
 
 # Add environment variables for ARM optimization
@@ -23,16 +21,20 @@ ENV OPENBLAS_CORETYPE=ARMV8
 ENV OPENBLAS_NUM_THREADS=1
 ENV OMP_NUM_THREADS=1
 
-# Sync dependencies using piwheels
-RUN --mount=type=cache,target=/root/.cache/uv \
-    uv sync --frozen --no-install-project --no-dev
+# Copy the ROOT pyproject.toml and uv.lock (workspace configuration)
+COPY pyproject.toml uv.lock ./
+COPY ./dashboard/pyproject.toml ./dashboard/
 
-COPY ./app ./app
-COPY ./main.py ./
-COPY ./.streamlit/ ./.streamlit/
+# Sync dependencies using the workspace configuration
+RUN --mount=type=cache,target=/root/.cache/uv \
+    uv sync --frozen --package dashboard --no-install-project --no-dev
+
+# Copy the dashboard package to the expected location
+COPY ./dashboard ./dashboard
+
 
 ENV PATH="/app/.venv/bin:$PATH"
-ENV PYTHONPATH="/app/collector"
+ENV PYTHONPATH="/app"
 ENV PYTHONUNBUFFERED=1
 
-CMD ["streamlit", "run", "main.py"]
+CMD ["streamlit", "run", "dashboard/main.py"]
