+++ dashboard/dashboard/config.py
