+++ dashboard/dashboard/core/database.py
@@ -1,7 +1,8 @@
-from app.core.settings import settings
 from sqlalchemy import MetaData, Table
 from sqlalchemy.engine import create_engine
 from sqlalchemy.orm import sessionmaker
+
+from dashboard.core.settings import settings
 
 engine = create_engine(settings.POSTGRES_URL, pool_size=10, max_overflow=20)
 SessionLocal = sessionmaker(bind=engine, autocommit=False, autoflush=False)
